Refactor GameSelectionIcon component
